Refactor: 프로파일 분리 및 패키지 수정
